Add globalClassName render option
This option adds a new per-render setting to allow a specific CSS class
to be added to the overall container #mcmap HTML element.

This element, and the extra class which is added, wraps the entire Leaflet
map making it possible to change global styles (on things like POI popups
and even map controls) per-render. For example, with custom CSS, it's
possible to make the controls dark for a nighttime render, but remain light
for a daytime render

Change-Id: I6689485b79c5735a7e1090a001b15d2eb5c77e72