Legend icons wrapped in a div
This change wraps the icons in the marker legend into a div, adds that
wrapper div to all markers even if there is no icon defined, and
transfers the styling to apply to the wrapper div instead.

This should have no impact on the way that the icon rendering is done in
the legend, but it does allow for easier extensibility in custom
stylesheets.

Change-Id: I7f6384694844daae3d41fc9a2e110cb9141d759e